remove hacks
