+++ src/Routers/PatientRouter.tsx
@@ -1,6 +1,5 @@
 import careConfig from "@careConfig";
 import { useRoutes } from "raviger";
-import { useEffect, useState } from "react";
 
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { AppSidebar, SidebarFor } from "@/components/ui/sidebar/app-sidebar";
@@ -9,6 +8,8 @@ import ErrorBoundary from "@/components/Common/ErrorBoundary";
 import ErrorPage from "@/components/ErrorPages/DefaultErrorPage";
 import { patientTabs } from "@/components/Patient/PatientDetailsTab";
 import { PatientHome } from "@/components/Patient/PatientHome";
+
+import useSidebarState from "@/hooks/useSidebarState";
 
 import PatientUserProvider from "@/Providers/PatientUserProvider";
 import { FacilitiesPage } from "@/pages/Facility/FacilitiesPage";
@@ -81,20 +82,10 @@ const AppointmentRoutes = {
 
 export default function PatientRouter() {
   const pages = useRoutes(DashboardRoutes);
-  const [sidebarOpen, setSidebarOpen] = useState(false);
 
   const appointmentPages = useRoutes(AppointmentRoutes);
 
-  // load the sidebar state from the cookie
-  useEffect(() => {
-    const cookie = document.cookie
-      .split("; ")
-      .find((row) => row.startsWith("sidebar:state"));
-    if (cookie) {
-      const value = cookie.split("=")[1];
-      setSidebarOpen(value === "true");
-    }
-  }, []);
+  const sidebarOpen = useSidebarState();
 
   if (!pages) {
     if (appointmentPages) {
